requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai>=1.10.0
python-dotenv==1.0.0
gunicorn==21.2.0
chromadb==0.4.22
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
sentence-transformers==2.2.2 